requirements.txt
pysnmp==4.4.12
pyasn1==0.4.8
psycopg2-binary
tabulate